Generally Finished Implementing the F4 Joker Detection Features
Using an implementation of a joker_created context from Winter, the current implementation correctly detects when a the relevant jokers are bought from the shop or generated (i.e. judgement) and can delete and generate the relevant jokers properly.

What is currently unknown is whether obtaining these jokers from a Buffoon Pack would maintain this functionality, which if it doesn't then may warrant further troubleshooting with the contexts or simply the creation of a blanket fix by forcing a check upon starting a blind.

Also Buffed the F4 Joker to +400 Chips and +4 xmult